+++ MbDCode/DirectionCosineIeqcJeqc.h
@@ -8,7 +8,7 @@ namespace MbD {
         //pAijIeJepEJ ppAijIeJepEIpEJ ppAijIeJepEJpEJ pAjOJepEJT ppAjOJepEJpEJ 
     public:
         DirectionCosineIeqcJeqc();
-        DirectionCosineIeqcJeqc(EndFrmcptr frmi, EndFrmcptr frmj, int axisi, int axisj);
+        DirectionCosineIeqcJeqc(EndFrmcptr frmi, EndFrmcptr frmj, size_t axisi, size_t axisj);
         void initialize();
         void initializeGlobally();
         void calcPostDynCorrectorIteration() override;
